[WIP] Pull out vertex code into it's own file
